Adds cross-compilation
Native fails to start successfully
