+++ build.sbt
@@ -14,8 +14,6 @@
  * limitations under the License.
  */
 
-name := "kitteh-redis"
-
 ThisBuild / baseVersion := "0.1"
 
 ThisBuild / organization := "com.codecommit"
@@ -24,22 +22,60 @@ ThisBuild / publishFullName := "Daniel Spiewak"
 
 ThisBuild / crossScalaVersions := Seq("2.13.7")
 
-enablePlugins(ScalaNativePlugin)
-
 val Fs2Version = "3.3.0"
 val Log4CatsVersion = "2.5.0"
 val Redis4CatsVersion = "1.0.0"
 
-libraryDependencies ++= Seq(
-  "org.typelevel" %% "cats-effect" % "3.3.14",
-  "org.typelevel" %% "log4cats-slf4j" % Log4CatsVersion,
-  "org.slf4j" % "slf4j-log4j12" % "1.7.9",
-  "org.scodec" %% "scodec-core" % "1.11.9",
-  "co.fs2" %% "fs2-io" % Fs2Version,
-  "co.fs2" %% "fs2-scodec" % Fs2Version,
-  "org.typelevel" %% "cats-effect-testing-specs2" % "1.4.0" % Test,
-  "org.typelevel" %% "log4cats-noop" % Log4CatsVersion % Test,
-  "org.specs2" %% "specs2-scalacheck" % "4.13.1" % Test,
-  "dev.profunktor" %% "redis4cats-effects" % Redis4CatsVersion % Test,
-  "dev.profunktor" %% "redis4cats-streams" % Redis4CatsVersion % Test
-)
+val isLinux = {
+  val osName = Option(System.getProperty("os.name"))
+  osName.exists(_.toLowerCase().contains("linux"))
+}
+val isMacOs = {
+  val osName = Option(System.getProperty("os.name"))
+  osName.exists(_.toLowerCase().contains("mac"))
+}
+
+lazy val `kitteh-redis` = crossProject(JVMPlatform, NativePlatform)
+  .in(file("."))
+  .settings(
+    name := "kitteh-redis",
+    libraryDependencies ++= Seq(
+      "org.typelevel" %%% "cats-effect" % "3.3.14",
+      "org.typelevel" %%% "log4cats-core" % Log4CatsVersion,
+      "org.scodec" %%% "scodec-core" % "1.11.10",
+      "co.fs2" %%% "fs2-io" % Fs2Version,
+      "co.fs2" %%% "fs2-scodec" % Fs2Version,
+      "org.typelevel" %% "cats-effect-testing-specs2" % "1.4.0" % Test,
+      "org.typelevel" %% "log4cats-noop" % Log4CatsVersion % Test,
+      "org.specs2" %% "specs2-scalacheck" % "4.13.1" % Test,
+      "dev.profunktor" %% "redis4cats-effects" % Redis4CatsVersion % Test,
+      "dev.profunktor" %% "redis4cats-streams" % Redis4CatsVersion % Test
+    ),
+    nativeConfig ~= { c =>
+      if (isLinux) { // brew-installed s2n
+        c.withLinkingOptions(
+          c.linkingOptions :+ "-L/home/linuxbrew/.linuxbrew/lib"
+        )
+      } else if (isMacOs) // brew-installed OpenSSL
+        c.withLinkingOptions(
+          c.linkingOptions :+ "-L/opt/homebrew/opt/openssl@1.1/lib"
+        )
+      else c
+    },
+    envVars ++= {
+      val ldLibPath =
+        if (isLinux)
+          Map("LD_LIBRARY_PATH" -> "/home/linuxbrew/.linuxbrew/lib")
+        else Map("LD_LIBRARY_PATH" -> "/usr/local/opt/openssl@1.1/lib")
+      Map("S2N_DONT_MLOCK" -> "1") ++ ldLibPath
+    }
+  )
+  .nativeSettings {
+    libraryDependencies ++= Seq(
+      "com.armanbilge" %%% "epollcat" % "0.1.0"
+    )
+  }
+
+lazy val root = project
+  .in(file("."))
+  .aggregate(`kitteh-redis`.jvm, `kitteh-redis`.native)
