+++ shared/src/main/scala/kitteh/Error.scala
