+++ shared/src/main/scala/kitteh/package.scala
