+++ shared/src/test/scala/CommandSpec.scala
