+++ shared/src/test/scala/EvalSpec.scala
